# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: mq API
  description: |
    Ethiopian artists platform API.

    Bounded contexts:
    - `content` — community wiki articles
    - `profile` — artist/institution presence
    - `art` — artwork posts
    - `onboarding` — admin vetting (admin routes)
  version: 0.1.0
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/artists/{slug}:
    get:
      tags: [profile]
      summary: Get artist profile by slug
      operationId: getArtistBySlug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Artist profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistProfile"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: mq_access_token
      description: httpOnly JWT session cookie set after OAuth or email/password login.
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ArtistProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        slug:
          type: string
        handle:
          type: string
        display_name:
          type: string
        bio:
          type: string
        born:
          type: string
        discipline:
          type: string
        tagline:
          type: string
        years_active:
          type: string
        portrait_url:
          type: string
        featured:
          type: boolean
        influences:
          type: array
          items:
            type: string
        in_residence:
          type: boolean
        residency_place:
          type: string
        open_for_commission:
          type: boolean
        contact:
          type: object
          properties:
            email: { type: string }
            phone: { type: string }
            website: { type: string }
            location: { type: string }
        social:
          type: object
          properties:
            instagram: { type: string }
            twitter: { type: string }
            telegram: { type: string }
        status:
          type: string
          enum: [draft, pending, approved]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...